Generate category pages in prebuild script
The script parses all notes (similar to the RSS feed generator script) in the fs, collects all existing categories from frontmatter and their associated posts, and generates a basic HTML page at /categories/name for each category. The page contains a list of posts tagged with that category (a post can be in multiple categories).