Render category links on a note header
The note header in the note template (page.html) is extended with categories if frontmatter for that note specifies them. Mustache doesn't support looping through arrays of strings, so to avoid having to add categories to frontmatter as an array of objects, the processNote build script uses regex to transform the categories string into links to category pages.