Use mustache templating instead of the build script
We 'loop' through categories using mustache and render one link for each category. The drawbacks are (1) that categories have to be added to frontmatter as an array of objects with the 'name' key, not a simple array, ad (2) that we can't add a comma between categories in a note header because mustache doesn't know what the last category is (to not but a comma after that one). Actually I could do this with CSS! Next commit.