Add categories (#41)
* Add categories to some notes' frontmatter

* Generate category pages in prebuild script

The script parses all notes (similar to the RSS feed generator script) in the fs, collects all existing categories from frontmatter and their associated posts, and generates a basic HTML page at /categories/name for each category. The page contains a list of posts tagged with that category (a post can be in multiple categories).

* Render category links on a note header

The note header in the note template (page.html) is extended with categories if frontmatter for that note specifies them. Mustache doesn't support looping through arrays of strings, so to avoid having to add categories to frontmatter as an array of objects, the processNote build script uses regex to transform the categories string into links to category pages.

* Use mustache templating instead of the build script

We 'loop' through categories using mustache and render one link for each category. The drawbacks are (1) that categories have to be added to frontmatter as an array of objects with the 'name' key, not a simple array, ad (2) that we can't add a comma between categories in a note header because mustache doesn't know what the last category is (to not but a comma after that one). Actually I could do this with CSS! Next commit.

* Add comma between categories with CSS

This also adds proper list markup to the categories. It was helpful for styling and actually improves semantics.